make confirm and setup interactors extend base one
