+++ fearless/Modules/Crowdloan/CrowdloanContributionConfirm/CrowdloanContributionConfirmViewFactory.swift
@@ -1,13 +1,19 @@
 import Foundation
+import SoraFoundation
 
 struct CrowdloanContributionConfirmViewFactory {
-    static func createView() -> CrowdloanContributionConfirmViewProtocol? {
+    static func createView(with _: ParaId, inputAmount _: Decimal) -> CrowdloanContributionConfirmViewProtocol? {
         let interactor = CrowdloanContributionConfirmInteractor()
         let wireframe = CrowdloanContributionConfirmWireframe()
 
         let presenter = CrowdloanContributionConfirmPresenter(interactor: interactor, wireframe: wireframe)
 
-        let view = CrowdloanContributionConfirmViewController(presenter: presenter)
+        let localizationManager = LocalizationManager.shared
+
+        let view = CrowdloanContributionConfirmVC(
+            presenter: presenter,
+            localizationManager: localizationManager
+        )
 
         presenter.view = view
         interactor.presenter = presenter
