+++ fearless/Modules/Crowdloan/CrowdloanContributionConfirm/CrowdloanContributionConfirmViewLayout.swift
@@ -1,0 +1,159 @@
+import UIKit
+
+final class CrowdloanContributionConfirmViewLayout: UIView {
+    let contentView: ScrollableContainerView = {
+        let view = ScrollableContainerView()
+        view.stackView.isLayoutMarginsRelativeArrangement = true
+        view.stackView.layoutMargins = UIEdgeInsets(top: 24.0, left: 0.0, bottom: 0.0, right: 0.0)
+        return view
+    }()
+
+    let accountView: DetailsTriangularedView = UIFactory.default.createAccountView()
+
+    let amountInputView: AmountInputView = {
+        let view = UIFactory().createAmountInputView(filled: true)
+        view.isUserInteractionEnabled = false
+        return view
+    }()
+
+    private(set) var estimatedRewardView: TitleValueView?
+
+    let leasingPeriodView = TitleMultiValueView()
+
+    let networkFeeConfirmView: NetworkFeeConfirmView = UIFactory().createNetworkFeeConfirmView()
+
+    var locale = Locale.current {
+        didSet {
+            if locale != oldValue {
+                applyLocalization()
+            }
+        }
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        backgroundColor = R.color.colorBlack()
+
+        setupLayout()
+        applyLocalization()
+    }
+
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    func bind(assetViewModel: AssetBalanceViewModelProtocol) {
+        amountInputView.priceText = assetViewModel.price
+
+        if let balance = assetViewModel.balance {
+            amountInputView.balanceText = R.string.localizable.commonAvailableFormat(
+                balance,
+                preferredLanguages: locale.rLanguages
+            )
+        } else {
+            amountInputView.balanceText = nil
+        }
+
+        amountInputView.assetIcon = assetViewModel.icon
+
+        let symbol = assetViewModel.symbol.uppercased()
+        amountInputView.symbol = symbol
+    }
+
+    func bind(feeViewModel: BalanceViewModelProtocol?) {
+        networkFeeConfirmView.networkFeeView.bind(viewModel: feeViewModel)
+    }
+
+    func bind(estimatedReward: String?) {
+        if let estimatedReward = estimatedReward {
+            createEstimatedRewardViewIfNeeded()
+            estimatedRewardView?.valueLabel.text = estimatedReward
+        } else {
+            removeEstimatedRewardViewIfNeeded()
+        }
+    }
+
+    private func applyLocalization() {
+        networkFeeConfirmView.locale = locale
+
+        leasingPeriodView.titleLabel.text = R.string.localizable.crowdloanLeasingPeriod(
+            preferredLanguages: locale.rLanguages
+        )
+
+        amountInputView.title = R.string.localizable
+            .walletSendAmountTitle(preferredLanguages: locale.rLanguages)
+
+        estimatedRewardView?.titleLabel.text = R.string.localizable.crowdloanReward(
+            preferredLanguages: locale.rLanguages
+        )
+    }
+
+    private func setupLayout() {
+        addSubview(contentView)
+        contentView.snp.makeConstraints { make in
+            make.top.equalTo(safeAreaLayoutGuide)
+            make.bottom.leading.trailing.equalToSuperview()
+        }
+
+        contentView.stackView.addArrangedSubview(accountView)
+        accountView.snp.makeConstraints { make in
+            make.width.equalTo(self).offset(-2.0 * UIConstants.horizontalInset)
+            make.height.equalTo(52)
+        }
+
+        contentView.stackView.setCustomSpacing(16.0, after: accountView)
+
+        contentView.stackView.addArrangedSubview(amountInputView)
+        amountInputView.snp.makeConstraints { make in
+            make.width.equalTo(self).offset(-2.0 * UIConstants.horizontalInset)
+            make.height.equalTo(72.0)
+        }
+
+        contentView.stackView.setCustomSpacing(16.0, after: amountInputView)
+
+        contentView.stackView.addArrangedSubview(leasingPeriodView)
+        leasingPeriodView.snp.makeConstraints { make in
+            make.width.equalTo(self).offset(-2.0 * UIConstants.horizontalInset)
+            make.height.equalTo(48.0)
+        }
+
+        addSubview(networkFeeConfirmView)
+
+        networkFeeConfirmView.snp.makeConstraints { make in
+            make.leading.bottom.trailing.equalToSuperview()
+        }
+    }
+
+    private func createEstimatedRewardViewIfNeeded() {
+        guard estimatedRewardView == nil else {
+            return
+        }
+
+        guard
+            let leasingPeriodIndex = contentView.stackView.arrangedSubviews.firstIndex(of: leasingPeriodView) else {
+            return
+        }
+
+        let view = TitleValueView()
+        view.titleLabel.text = R.string.localizable.crowdloanReward(preferredLanguages: locale.rLanguages)
+
+        contentView.stackView.insertArrangedSubview(view, at: leasingPeriodIndex + 1)
+        view.snp.makeConstraints { make in
+            make.width.equalTo(self).offset(-2.0 * UIConstants.horizontalInset)
+            make.height.equalTo(48.0)
+        }
+
+        estimatedRewardView = view
+    }
+
+    private func removeEstimatedRewardViewIfNeeded() {
+        guard let estimatedRewardView = estimatedRewardView else {
+            return
+        }
+
+        contentView.stackView.removeArrangedSubview(estimatedRewardView)
+        estimatedRewardView.removeFromSuperview()
+    }
+}
