+++ fearless/Modules/Crowdloan/CrowdloanContributionSetup/CrowdloanContributionSetupProtocols.swift
@@ -18,24 +18,11 @@ protocol CrowdloanContributionSetupPresenterProtocol: AnyObject {
     func proceed()
 }
 
-protocol CrowdloanContributionSetupInteractorInputProtocol: AnyObject {
-    func setup()
-    func estimateFee(for amount: BigUInt)
-}
+protocol CrowdloanContributionSetupInteractorInputProtocol: CrowdloanContributionInteractorInputProtocol {}
 
-protocol CrowdloanContributionSetupInteractorOutputProtocol: AnyObject {
-    func didReceiveCrowdloan(result: Result<Crowdloan, Error>)
-    func didReceiveDisplayInfo(result: Result<CrowdloanDisplayInfo?, Error>)
-    func didReceiveAccountInfo(result: Result<AccountInfo?, Error>)
-    func didReceiveBlockNumber(result: Result<BlockNumber?, Error>)
-    func didReceiveBlockDuration(result: Result<BlockTime, Error>)
-    func didReceiveLeasingPeriod(result: Result<LeasingPeriod, Error>)
-    func didReceiveMinimumBalance(result: Result<BigUInt, Error>)
-    func didReceivePriceData(result: Result<PriceData?, Error>)
-    func didReceiveFee(result: Result<RuntimeDispatchInfo, Error>)
-}
+protocol CrowdloanContributionSetupInteractorOutputProtocol: CrowdloanContributionInteractorOutputProtocol {}
 
 protocol CrowdloanContributionSetupWireframeProtocol: AlertPresentable, ErrorPresentable,
     CrowdloanErrorPresentable {
-    func showConfirmation(from view: CrowdloanContributionSetupViewProtocol?, inputAmount: Decimal)
+    func showConfirmation(from view: CrowdloanContributionSetupViewProtocol?, paraId: ParaId, inputAmount: Decimal)
 }
