+++ fearless/Modules/Crowdloan/CrowdloanContributionSetup/CrowdloanContributionSetupWireframe.swift
@@ -1,7 +1,14 @@
 import Foundation
 
 final class CrowdloanContributionSetupWireframe: CrowdloanContributionSetupWireframeProtocol {
-    func showConfirmation(from _: CrowdloanContributionSetupViewProtocol?, inputAmount _: Decimal) {
-        // TODO: FLW-861
+    func showConfirmation(from view: CrowdloanContributionSetupViewProtocol?, paraId: ParaId, inputAmount: Decimal) {
+        guard let confirmationView = CrowdloanContributionConfirmViewFactory.createView(
+            with: paraId,
+            inputAmount: inputAmount
+        ) else {
+            return
+        }
+
+        view?.controller.navigationController?.pushViewController(confirmationView.controller, animated: true)
     }
 }
