--- fearless/Modules/CrowdloanContributionConfirm/CrowdloanContributionConfirmViewLayout.swift
@@ -1,12 +1,0 @@
-import UIKit
-
-final class CrowdloanContributionConfirmViewLayout: UIView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-    }
-
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
